refactor: extract model list
